Extract and export the `nullable` serializer
