+++ packages/umi-serializer-beet/src/index.ts
@@ -3,6 +3,7 @@ export * from './createBeetSerializer';
 export * from './bytes';
 export * from './errors';
 export * from './map';
+export * from './nullable';
 export * from './option';
 export * from './set';
 export * from './tuple';
